requirements.txt
cython
six >= 1.5
numpy >= 1.10
scipy >= 0.16
astropy >= 1.2.1
matplotlib>=1.4.1
lscsoft-glue>=1.55.2
